Small bug fixes, updated formatting, file structure
