+++ tabs/geometry/frames/landing_gear_frame.py
@@ -1,12 +1,13 @@
 from PyQt6.QtWidgets import QWidget, QLabel, QVBoxLayout
 
-from color import Color
+from widgets.color import Color
 
-class WingsFrame(QWidget):
+
+class LandingGearFrame(QWidget):
     def __init__(self):
-        super(WingsFrame, self).__init__()
+        super(LandingGearFrame, self).__init__()
 
         layout = QVBoxLayout()
-        layout.addWidget(QLabel("Wings Frame"))
+        layout.addWidget(QLabel("Landing Gear Frame"))
         layout.addWidget(Color("lightblue"))
         self.setLayout(layout)
