+++ tabs/geometry/geometry.py
@@ -1,15 +1,13 @@
 from PyQt6.QtCore import Qt
-from PyQt6.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout, QComboBox
+from PyQt6.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout, QComboBox, QLayout
 
-
-
+from tabs.geometry.frames.default_frame import DefaultFrame
+from tabs.geometry.frames.energy_network_frame import EnergyNetworkFrame
+from tabs.geometry.frames.fuselage_frame import FuselageFrame
+from tabs.geometry.frames.landing_gear_frame import LandingGearFrame
+from tabs.geometry.frames.nacelle_frame import NacellesFrame
+from tabs.geometry.frames.wings_frame import WingsFrame
 from widgets.color import Color
-from tabs.frames.default_frame import DefaultFrame
-from tabs.frames.wings_frame import WingsFrame
-from tabs.frames.fuselage_frame import FuselageFrame
-from tabs.frames.nacelle_frame import NacellesFrame
-from tabs.frames.landing_gear_frame import LandingGearFrame
-from tabs.frames.energy_network_frame import EnergyNetworkFrame
 
 
 class GeometryWidget(QWidget):
@@ -23,18 +21,19 @@ class GeometryWidget(QWidget):
         self.main_extra_frame = None  # Initialize as None
 
         self.tree_frame_layout = QVBoxLayout(self.tree_frame)
-        
+
         # Set a background color for tree_frame
         tree_frame_style = """
             background-color: navy
         """
         self.tree_frame.setStyleSheet(tree_frame_style)
-        
+
         # Create a QComboBox and add options
         self.dropdown = QComboBox()
-        options = ["Select an option", "Add Fuselage", "Add Wings", "Add Nacelles", "Add Landing Gear", "Add Energy Network"]
+        options = ["Select an option", "Add Fuselage", "Add Wings", "Add Nacelles", "Add Landing Gear",
+                   "Add Energy Network"]
         self.dropdown.addItems(options)
-        
+
         # Style the dropdown with a colored background
         dropdown_style = """
             QComboBox {
@@ -49,7 +48,7 @@ class GeometryWidget(QWidget):
             }
         """
         self.dropdown.setStyleSheet(dropdown_style)
-        
+
         self.tree_frame_layout.addWidget(self.dropdown, alignment=Qt.AlignmentFlag.AlignTop)
 
         main_layout.addWidget(Color("navy"), 7)
@@ -70,33 +69,29 @@ class GeometryWidget(QWidget):
         frames = [DefaultFrame, FuselageFrame, WingsFrame, NacellesFrame, LandingGearFrame, EnergyNetworkFrame]
 
         if 0 <= index < len(frames):
-            # Replace main_extra_frame with the selected frame
+            # Replace main_extra_frame with the selected frames
             new_frame = frames[index]()
             self.replace_main_extra_frame(new_frame)
 
     def replace_main_extra_frame(self, new_frame):
         old_frame = self.main_extra_frame
-    
+
         if old_frame is not None:
             main_layout = self.layout().itemAt(1)  # Assuming main_extra_frame is at index 1
-    
+
+            assert isinstance(main_layout, QLayout)
             if main_layout is not None:
-                # Remove the old frame from the layout
+                # Remove the old frames from the layout
                 main_layout.removeWidget(old_frame)
                 old_frame.setParent(None)
-    
-        # Add the new frame to the layout
+
+        # Add the new frames to the layout
         main_layout = self.layout().itemAt(1)  # Assuming main_extra_frame is at index 1
+        assert isinstance(main_layout, QLayout)
         main_layout.addWidget(new_frame, 3)
-    
+
         # Update the reference to the current main_extra_frame
         self.main_extra_frame = new_frame
-
-
-
-
-
-
 
 
 def get_widget() -> QWidget:
